Update suggestion_109.html
The best way to eat better? Avoid temptation altogether! Here is a list of items  to avoid at the grocery store with hidden  copious amounts of calories!

1. Canned tomatoes

>Expandable:
An endocrinologist and expert on the topic of the synthetic oestrogen bisphenol-A (BCA), linked to heart disease and infertility, won’t go near canned tomatoes. Tin cans are lined with a resin containing BCA which is especially a problem with canning tomatoes, as the acid in tomato breaks this down in dangerous amounts. This is a serious health concern for everyone who loves a Spag Bol, especially children. My advice: if you still want the convenience of stored, ready-to-cook tomatoes, opt for sauces and passata in glass bottles.

2. Microwave popcorn:

>Expandable:
Another poisonous packaging issue: Perfluorooctanoic acid (PFOA) lines the bags of those popcorn bags, and the heat in the microwave leaches this straight onto your movie munchies. The UCLA links this compound to infertility. My advice: Corn kernels + butter + sea salt + plus a big pot (with a lid!) Simple.

3. Conventional Apples

>Expandable:
There’s no coincidence that farm workers have higher rates of many cancers. Of all common fruits, apples are the most heavily and frequently doused with pesticides. Pesticide reside on conventional fruits is also linked to Parkinson’s. To limit exposure, be wary of apples especially. My advice: Organic. Or at the very least, wash and peel.