Update suggestion_110.html
Food just ain’t what it used to be!  Organic produce is labeled organic when it has been grown, raised, harvested and packaged without the use of harmful chemicals, such as fertilizers, pesticides, insecticides, growth hormones or antibiotics. Organic also means the produce has not been genetically modified. When we choose organic we not only look after the health of our own body and immune system but also the health of our family, the health of our soil and the health of our planet. So everyone wins. The best way to get organic produce in the home is to grow it yourself. If that's not an option, try your local grocer, farmers market, or search the internet for a company that delivers organic produce to your door.

To find local produce in your area,  (Canada) check out: www.cog.ca