Update suggestion_116.html
Spiritual wellness is the process of getting in touch with your  spiritual core. It is the process of looking within to become more in touch with the spiritual dimensions of your life . 

An attitude of active searching  increases your options and your potential for  spiritual centering. Don’t shut doors before you check out what’s behind them .