Update suggestion_120.html
Feel the greatness

This is a great day to be alive. This is a great day to be who you are, where you are, and the way you are.

On this day, you can begin by assuming the best. Then you can follow through and do everything in your power to make it happen.

Today there will be all kinds of challenges and frustrations. And today, you can use each one of them as a way to grow stronger, more compassionate, more capable and more accomplished.
This is a day that’s too unique and precious to waste. This is a day that’s filled with new and exciting possibilities for making a difference.

Though there are certainly things to complain about, there is much more to be thankful for. Focus on the gratitude instead of the shortcomings, and you’ll invite today’s special flavor of abundance into your life.

This is a great day to be positive, purposeful and filled with enthusiasm for all you can do. Feel all the potential greatness in this day, and get busy bringing it to life in your own special way.


Read more at http://greatday.com/motivate/140402.html#eM8GgVgo4ZmbFtwO.99