Update suggestion_121.html
Start right away

Careful and detailed planning is essential to any achievement, and yet the planning is not the achievement. To achieve any goal you must act on it, and the sooner you begin to act, the better.
The first thing to do is to do something. Instead of speculating, or debating, or making excuses, or wondering, step forward and take action.

You cannot learn what works until you actually do some work. Begin, learn, adjust, and continue acting with growing wisdom and experience until the goal is reached.
Yes, you can benefit greatly from the experience of others. And the way to do that is to continue adding your own experience to it.

If there’s something you intend to do, get started. If you don’t take the first step immediately, it’s not really an intention but merely an empty wish.

You are fully capable and fully deserving of the great, rewarding experience of achievement. Start on it right away, right now, and bring that experience to life.


Read more at http://greatday.com/motivate/140319.html#WST46VATKr3dCxGZ.99